+++ src/main/java/bowling/domain/BallThrow.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 import java.util.Objects;
 import java.util.Optional;
