+++ src/main/java/bowling/domain/IllegalBallThrownException.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 class IllegalBallThrownException extends RuntimeException {
 }
