+++ src/main/java/bowling/domain/IllegalFallingPinsException.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 class IllegalFallingPinsException extends RuntimeException {
 }
