+++ src/test/java/bowling/domain/BallThrowTest.java
@@ -1,5 +1,6 @@
-package bowling;
+package bowling.domain;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -18,7 +19,7 @@ import static org.assertj.core.api.Assertions.*;
 public class BallThrowTest {
     @Test
     void createAndEquals() {
-        assertThat(new BallThrow(0)).isEqualTo(new BallThrow(0));
+        Assertions.assertThat(new BallThrow(0)).isEqualTo(new BallThrow(0));
     }
 
     @DisplayName("10개 초과의 핀을 쓰러뜨릴 수 없다")
