+++ src/test/java/bowling/domain/FrameTest.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -8,8 +8,9 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.Optional;
 
-import static bowling.Scoring.*;
-import static org.assertj.core.api.Assertions.*;
+import static bowling.domain.Scoring.STRIKE;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
  * #### 프레임
